wrapped up converting TrailDetailsSideIconMenu to typescript
