+++ client/src/main.tsx
@@ -9,7 +9,6 @@ import "@fontsource/roboto/700.css";
 import { BrowserRouter } from "react-router-dom";
 
 
-
 const container = document.getElementById('root') as HTMLElement;
 ReactDOM.createRoot(container).render(
   <React.Fragment>
