+++ server/package.json
@@ -17,6 +17,10 @@
     "pg": "^8.10.0"
   },
   "devDependencies": {
+    "@types/cookie-parser": "^1.4.3",
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.16.0",
     "nodemon": "^2.0.22"
   }
 }
