+++ tsconfig.node.json
@@ -3,10 +3,10 @@
     "composite": true,
     "module": "ESNext",
     "moduleResolution": "Node",
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
   },
   "include": [
     "vite.config.ts",
-    "./src"
+    "./src",
   ]
 }
